+++ lib/src/types/big_int.dart
@@ -1,4 +1,3 @@
-import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer/dart/element/nullability_suffix.dart';
 import 'package:analyzer/dart/element/type.dart';
 import 'package:args_generator/src/types/type_helper.dart';
@@ -18,16 +17,16 @@ class TypeHelperBigInt extends TypeHelper {
   @override
   bool matchesType(DartType type) {
     if (type is InterfaceType) {
-      final element = type.element;
+      final element = type.element3;
 
-      return element.name == 'BigInt';
+      return element.name3 == 'BigInt';
     }
     return false;
   }
 
   /// Decodes a `BigInt` value from the provided arguments.
   ///
-  /// - [field]: The [ParameterElement] representing the field to decode.
+  /// - [field]: The field to decode.
   /// - [defaultValue]: The default value for the field, if any.
   ///
   /// Converts the value in the `args` map (with the field name converted
@@ -37,7 +36,7 @@ class TypeHelperBigInt extends TypeHelper {
   /// Returns:
   /// A string representing the code to decode the `BigInt` value.
   @override
-  String decode(ParameterElement field, String? defaultValue) {
+  String decode(ArgField field, String? defaultValue) {
     final key = field.name.convertToKebabCase();
     final arg = "args['$key']";
     final isNullable = field.type.nullabilitySuffix != NullabilitySuffix.none;
@@ -51,7 +50,7 @@ class TypeHelperBigInt extends TypeHelper {
 
   /// Encodes a `BigInt` field into a map format.
   ///
-  /// - [field]: The [ParameterElement] representing the field to encode.
+  /// - [field]: The field to encode.
   ///
   /// If the field is nullable, the generated code includes a conditional check
   /// to ensure the field is not `null` before adding it to the map. The field's
@@ -60,7 +59,7 @@ class TypeHelperBigInt extends TypeHelper {
   /// Returns:
   /// A string representing the code to encode the `BigInt` value.
   @override
-  String encode(ParameterElement field) {
+  String encode(ArgField field) {
     final isNullable = field.type.nullabilitySuffix != NullabilitySuffix.none;
     final name = field.name;
     final key = name.convertToKebabCase();