+++ lib/src/types/bool.dart
@@ -1,4 +1,3 @@
-import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer/dart/element/nullability_suffix.dart';
 import 'package:analyzer/dart/element/type.dart';
 import 'package:args_generator/src/types/type_helper.dart';
@@ -20,7 +19,7 @@ class TypeHelperBool extends TypeHelper {
 
   /// Decodes a `bool` value from the provided arguments.
   ///
-  /// - [field]: The [ParameterElement] representing the field to decode.
+  /// - [field]: The field to decode.
   /// - [defaultValue]: The default value for the field, if any.
   ///
   /// Checks if the key (derived from the field name converted to kebab-case)
@@ -32,7 +31,7 @@ class TypeHelperBool extends TypeHelper {
   /// Returns:
   /// A string representing the code to decode the `bool` value.
   @override
-  String decode(ParameterElement field, String? defaultValue) {
+  String decode(ArgField field, String? defaultValue) {
     final key = field.name.convertToKebabCase();
     final isNullable = field.type.nullabilitySuffix != NullabilitySuffix.none;
     return '''
@@ -44,7 +43,7 @@ args.containsKey('$key')
 
   /// Encodes a `bool` field into a map format.
   ///
-  /// - [field]: The [ParameterElement] representing the field to encode.
+  /// - [field]: The field to encode.
   ///
   /// If the field is nullable, the generated code includes a conditional check
   /// to ensure the field is not `null` before adding it to the map. The field's
@@ -53,7 +52,7 @@ args.containsKey('$key')
   /// Returns:
   /// A string representing the code to encode the `bool` value.
   @override
-  String encode(ParameterElement field) {
+  String encode(ArgField field) {
     final key = field.name.convertToKebabCase();
     final isNullable = field.type.nullabilitySuffix != NullabilitySuffix.none;
     final name = field.name;