+++ lib/src/types/iterable.dart
@@ -1,4 +1,3 @@
-import 'package:analyzer/dart/element/element.dart';
 import 'package:analyzer/dart/element/nullability_suffix.dart';
 import 'package:analyzer/dart/element/type.dart';
 import 'package:args_generator/src/types/type_helper.dart';
@@ -18,16 +17,16 @@ class TypeHelperIterable extends TypeHelper {
   @override
   bool matchesType(DartType type) {
     if (type is InterfaceType) {
-      final element = type.element;
+      final element = type.element3;
 
-      return element.name == 'List' || element.name == 'Iterable';
+      return element.name3 == 'List' || element.name3 == 'Iterable';
     }
     return false;
   }
 
   /// Decodes an `Iterable` value from the provided arguments.
   ///
-  /// - [field]: The [ParameterElement] representing the field to decode.
+  /// - [field]: The field to decode.
   /// - [defaultValue]: The default value for the field, if any (not used here).
   ///
   /// The method checks if the field's name (converted to kebab-case) exists in
@@ -39,7 +38,7 @@ class TypeHelperIterable extends TypeHelper {
   /// Returns:
   /// A string representing the code to decode the `Iterable` value.
   @override
-  String decode(ParameterElement field, String? defaultValue) {
+  String decode(ArgField field, String? defaultValue) {
     final key = field.name.convertToKebabCase();
     final isNullable = field.type.nullabilitySuffix != NullabilitySuffix.none;
 
@@ -55,7 +54,7 @@ args.containsKey('$key')
 
   /// Encodes an `Iterable` field into a map format.
   ///
-  /// - [field]: The [ParameterElement] representing the field to encode.
+  /// - [field]: The field to encode.
   ///
   /// If the field is nullable, the generated code includes a conditional check
   /// to ensure the field is not `null` before adding it to the map. The field's
@@ -65,7 +64,7 @@ args.containsKey('$key')
   /// Returns:
   /// A string representing the code to encode the `Iterable` value.
   @override
-  String encode(ParameterElement field) {
+  String encode(ArgField field) {
     final key = field.name.convertToKebabCase();
     final isNullable = field.type.nullabilitySuffix != NullabilitySuffix.none;
     final name = field.name;